requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
datasets==2.12.0
decorator==5.2.1
dill==0.3.6
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.13.1
frozenlist==1.7.0
fsspec==2022.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.44.0
MarkupSafe==2.1.5
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.14
networkx==3.3
numba==0.61.2
numpy==1.24.0
openai==1.97.1
openai-whisper==20250625
packaging==25.0
pandas==2.3.1
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
psutil==7.0.0
pyarrow==12.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
responses==0.18.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.53.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
